+++ src/main/kotlin/alvarx4z/medialibrary/domain/shared/professional/Date.kt
@@ -1,4 +1,4 @@
-package alvarx4z.medialibrary.domain.shared.professional.birth
+package alvarx4z.medialibrary.domain.shared.professional
 
 import alvarx4z.medialibrary.domain.shared.invariant.PastDate
 import java.time.LocalDate
