+++ src/test/kotlin/alvarx4z/medialibrary/domain/shared/professional/DateTest.kt
@@ -1,4 +1,4 @@
-package alvarx4z.medialibrary.domain.shared.professional.birth
+package alvarx4z.medialibrary.domain.shared.professional
 
 import alvarx4z.medialibrary.SharedHelper
 import java.time.Month
